+++ metagenomescope/graph/assembly_graph.py
@@ -66,7 +66,7 @@ class AssemblyGraph(object):
     https://www.thedigitalcatonline.com/blog/2014/08/20/python-3-oop-part-3-delegation-composition-and-inheritance/
     """
 
-    def __init__(self, graph_fp, agp_fp=None, flye_info_fp=None):
+    def __init__(self, graph_fp, agp_fp=None, gaf_fp=None, flye_info_fp=None):
         """Parses the input graph file and initializes the AssemblyGraph.
 
         Parameters
@@ -77,6 +77,10 @@ class AssemblyGraph(object):
         agp_fp: str or None
             If specified, this should be a path to an AGP file describing paths
             in the graph.
+
+        gaf_fp: str or None
+            If specified, this should be a path to a Rukki-/Verkko-style GAF
+            file describing paths in the graph.
 
         flye_info_fp: str or None
             If specified, this should be a path to a Flye assembly_info.txt
@@ -96,6 +100,8 @@ class AssemblyGraph(object):
         self.basename = misc_utils.get_basename_if_fp(graph_fp)
         self.agp_filename = agp_fp
         self.agp_basename = misc_utils.get_basename_if_fp(agp_fp)
+        self.gaf_filename = gaf_fp
+        self.gaf_basename = misc_utils.get_basename_if_fp(gaf_fp)
         self.flye_info_filename = flye_info_fp
         self.flye_info_basename = misc_utils.get_basename_if_fp(flye_info_fp)
 
@@ -344,6 +350,17 @@ class AssemblyGraph(object):
                 f"{ui_utils.pluralize(len(agp_paths), 'path')}."
             )
             path_utils.merge_paths(input_paths, agp_paths)
+
+        if self.gaf_filename is not None:
+            logger.debug(f'  Loading input GAF file "{self.gaf_basename}"...')
+            gaf_paths = path_utils.get_paths_from_gaf(
+                self.gaf_filename, self.orientation_in_name
+            )
+            logger.debug(
+                "  ...Done. It contained "
+                f"{ui_utils.pluralize(len(gaf_paths), 'path')}."
+            )
+            path_utils.merge_paths(input_paths, gaf_paths)
 
         if self.flye_info_filename is not None and self.filetype == "DOT":
             logger.debug(